data-scraper/app/config/settings.py
import os
from dotenv import load_dotenv

env_file = ".env"
load_dotenv(env_file)

MONGO_URI = os.getenv("MONGO_URI")

MYSQL_HOST = os.getenv("MYSQL_HOST")
MYSQL_PORT = os.getenv("MYSQL_PORT")
MYSQL_USER = os.getenv("MYSQL_USER")
MYSQL_PASSWORD = os.getenv("MYSQL_PASSWORD")
MYSQL_DATABASE = os.getenv("MYSQL_DATABASE")
