data-scraper/pyproject.toml
[tool.poetry]
name = "wanted-scraper"
version = "0.1.0"
description = ""
authors = ["Ruo Lee <comsa33@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.6.3"
icecream = "^2.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
fastapi = "^0.110.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
